FIXED FIXED FIXED FIXED FIXED

+++ internal/util/log.go
@@ -9,6 +9,9 @@ var Logger *log.Logger
 
 func init() {
 	// Clear log file
+	if os.IsNotExist(os.MkdirAll("C:\\logs", 0755)) {
+		panic("Unable to create logs directory")
+	}
 	os.Remove("C:\\logs\\debug.log")
 	// Create log file
 	f, err := os.OpenFile("C:\\logs\\debug.log", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
